Input form rendering, changed states on Form.js and App.js, still not functional
